added drag and drop feature from react-clickdrag. fixed the indexing: new Notes had duplicate index id cuz we relied on the array length, which shrinks if we remove Notes. Kept the ClickDragExample as reference but has no use. Hardest part was updating the this.style in NoteDraggable.js L:52
