+++ package.json
@@ -4,6 +4,7 @@
   "private": true,
   "dependencies": {
     "react": "^16.5.1",
+    "react-clickdrag": "^4.0.0",
     "react-dom": "^16.5.1",
     "react-icons": "^3.0.5",
     "react-scripts": "1.1.5"
